Minor Refactoring
