+++ ui/src/components/Tabs/AdminTab/ContentEditor/FileRemover.js
@@ -1,7 +1,7 @@
 import React, { useEffect, useState } from 'react';
 import {Form, Button, Accordion} from 'semantic-ui-react';
-import { config } from '../../util/functions/constants';
-import { SecureFetch } from '../../util/functions/secureFetch';
+import { config } from '../../../util/functions/constants';
+import { SecureFetch } from '../../../util/functions/secureFetch';
 
 /*
 * I need to pass in file paths from file editor to here through props.
