+++ ui/src/components/Tabs/AdminTab/ContentEditor/OverviewEditor.js
@@ -1,7 +1,7 @@
 import React, { useEffect, useState } from 'react';
 import {Form, Button, Accordion} from 'semantic-ui-react';
-import { config } from '../../util/functions/constants';
-import { SecureFetch } from '../../util/functions/secureFetch';
+import { config } from '../../../util/functions/constants';
+import { SecureFetch } from '../../../util/functions/secureFetch';
 
 const PAGE = {OVERVIEW: "Overview", SPONSOR: "Sponsor"};
